requirements.txt
absl-py==0.8.0
astor==0.8.0
dronekit==2.9.2
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.1
Markdown==3.1.1
monotonic==1.5
numpy==1.17.1
opencv-python==4.1.1.26
Pillow==6.1.0
pkg-resources==0.0.0
protobuf==3.9.1
pycocotools==2.0.0
pymavlink==2.3.8
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
